Rudimentary IAPD file import
The essentials are in place and working: the one test file
I have is read correctly and all relevant data is extracted.
Work remaining to be done is mainly UI-related: choosing
file type on import (currently has to be suffixed .iapd),
selecting treatment type, informative errors for malformed
files, etc.