More work on VGP calculations
Site location is now stored in the Site object; VGP is calculated and
stored in Site whenever a Fisher/GC direction is determined. Location
is saved and read to/from PuffinPlot files. Location and VGP are exported
in site data export. As yet, location cannot be set by user, and
location and VGP are not displayed anywhere.

Related to: a88/7b9 Implement VGP calculation