Safer a95 plotting, etc.
Some plots were assuming that a95 will always be a valid number.
This is not guaranteed. Plotting functions now check for validity
rather than blithely attempting to construct a circle with a
radius of NaN.

Also fixed a bug in Plot which caused clicks to be handled even on
currently invisible plots, with potentially confusing results for the
user.