Support dynamic Jython loading; remove Jython jar from buil
Issue UUID: a88/0df 0dfc5e78-620d-4e23-820f-ad062997fce2

Jython is now an optional extra for PuffinPlot. If Jython is required
by the user, PuffinPlot downloads the jar, caches it locally, and
loads it dynamically.

Dynamic Jython support is implemented in this commit, but only for Python
scripts run via the GUI. Command-line Python scripting is broken, as is
Python support in the great-circle validity testing method. (The latter
will probably be replaced with JavaScript.)