Ivy now handles almost all dependencies.
fetch-libs.sh has been removed, and most jars are now fetched
automatically by Ivy during the Ant build process. The remaining
special cases are the Apple Java extensions (currently checked
into the repository) which should probably be replaced with the
Orange extensions, and the JAMA library which is included as
source code and could be replaced either with the jar-packaged
version or (perhaps better) with commons-math3 (already in use).