Move RPI functionality to new RpiDataset class
Previously RPI calculation was in PuffinApp, which was always intended
as a temporary home. It's not in its own class, RpiDataset, and
calculation has been separated from writing results to a file. The
code's a bit rough at the moment, but output has been tested against
the previous implementation in PuffinApp to ensure that results are
identical.