Add RpiDataset.calculateWithMagSus method
This is a quick and dirty implementation which performs the calculation
correctly, but uses RpiDatum in an ugly way: RpiDatum is currently set
up for ARM-derived data only, so putting MS-derived data in it abuses
the semantics a bit. There's no GUI hook for it at present, but it's
usable (and tested) as is from Jython.