Add test for Vec3.correctForm
Vec3 now has virtually full unit test coverage. JaCoCo reports 97%
coverage, but most of the missing 3% is on asserts. Some of these are
asserting parameter validity and and will be replaced with explicit,
exception-throwing checks. Most of the remainder are internal asserts
of mainly documentary value, asserting against "can't-happen" cases,
and can't practically be made to fail.