Preparations for migration to git
 - Add tasks to the build file's -pre-jar target to extract revision
   information from a git repo and write it to the build.properties
   file.

 - Add a fromGitProperties static factory method to Version to create a
   Version object from the git information in build.properties.

At present, there is no git repo so the information written to
build.properties is invalid, and PuffinApp still constructs its version
data using the fromMercurialProperties method. However, the pieces are
now in place, so when the repository is migrated to git a one-line
change in PuffinApp should be sufficient to update the construction
of PuffinApp.version accordingly.