Refactor ExceptionHandler into a top-level class
Extract ExceptionHandler from PuffinApp and make it a top-level class
in the puffinplot package. Add a setApp method to associate it with
a PuffinPlot instance (it can also work without one). Usage is now as
follows when PuffinPlot is started as a GUI application:

1. instantiate ExceptionHandler and set it as the uncaught
   exception handler to catch any exceptions thrown during PuffinApp
   instantiation.

2. Instantiate PuffinApp.

3. Supply the new PuffinApp to the ExceptionHandler via setApp, which
   gives it access to log data, etc. for crash reports.

Additionally: change PuffinApp's default log level from FINE to INFO.