Add JUnit 4.12 dependency to ivy.xml
Add an explicit dependency on JUnit 4.12 to ivy.xml. This isn't
necessary for NetBeans, but makes it easier to import the project into
IntelliJ IDEA. Without this dependency, IDEA will try to run
PuffinPlot's unit tests using JUnit 4.10, which is pulled in as a
transitive dependency of the FreeHEP libraries. This fails because some
of the tests use assertNotEquals, which isn't present in JUnit 4.10.
Adding the dependency directly causes IDEA to use JUnit 4.12 for the
unit tests instead.