Rename the Datum class to TreatmentStep
"Datum" was an ambiguous and poorly chosen class name which has
persisted since the earliest versions of PuffinPlot. I am finally
renaming it to the far more descriptive "TreatmentStep". Because Datum /
TreatmentStep is PuffinPlot's fundamental data class, this refactoring
involves a large number of changes. It was performed in IntelliJ IDEA
2018.3.3 (Community Edition) Build #IC-183.5153.38, since NetBeans
couldn't handle it. As well as the class name itself, IDEA has changed
identifiable occurrences of "datum" in variable names, comments, the
manual, strings, tests, getters, and setters. There is still more work
to do (various identifiers involving the term "data", many occurrences
of "TreatmentStep d", and the classes DatumField, DatumMomentComparator,
and DatumTreatmentComparator), but this is an excellent start. I have
skimmed the refactored code, run the test suite, and tried out the
application, and am confident that no bugs have been introduced.