GraphDisplay: index plots by class, not name
GraphDisplay maintained a map from plot class names (as Strings) to plot
instances. I have changed this map to use the plots' Class objects
rather than class names as keys, as this is safer: there's less chance
of triggering a runtime error due to a typo or a class rename. I made
the corresponding changes to PuffinApp and GraphDisplay's subclasses to
accommodate the change (mostly in MainGraphDisplay), and did some minor
housekeeping where needed (adding missing Javadoc and reformatting
code).