Redo plot constructors to remove Swing dependency
Previously, plot constructors took three arguments: a GraphDisplay, a
PlotParams, and a Preferences. The GraphDisplay was in fact only used to
calculate a scaling factor based on its maximum size, but meant that
every plot depended on a Swing component. In the new architecture, the
plot constructors only take a PlotParams. The PlotParams interface has
been extended with methods that provide the scaling factor and the
Preferences object. This decouples the plots from the Swing GUI, making
non-interactive plot construction (e.g. for unit testing or scripting)
easier.