Clean up code for some plot classes; minor bug fix
General clean-up of some more plot classes -- mainly code style and
formatting. Also found a minor bug in DepthPlot: colour was not being
set explicitly before drawing axes; generally this is black, so it
didn't matter, but if the NRM histogram is active it happens to be grey,
so the axes were being drawn in grey. Now fixed.