Remove TreatmentStep#getTreatmentStep, etc.
Remove the redundant and confusingly named
TreatmentStep#getTreatmentStep: it did the same thing as
getTreatmentLevel, with which I replaced it in the few places it was
being called. I also renamed Sample#getLastGcStep to getLastGcStepLevel
and getFirstGcStep to getFirstGcStepLevel, and did the usual code style
and formatting improvements and Javadoc tidy-ups where needed.

rename getLastGcStepLevel etc.