FisherParams#isA95Valid() default implementation
Now that PuffinPlot is running on Java 8, it's possible to have default
method implementations in interfaces. Previously, isA95Valid() was
defined in FisherParams and implemented separately, but identically, in
FisherValues and GreatCircles. I have now added a default implementation
of isA95Valid() to FisherParams instead, and removed the implementations
in FisherValues and GreatCircles.