Remove unneeded Preferences arguments, etc.
The constructors of SiteMeanDisplay, SiteMeanWindow, and
SuiteEqAreaDisplay took Preferences objects as an argument, but none of
them made any use of it; I have removed it from the constructor
arguments in all three classes, along with an unused local Preferences
variable in MainGraphDisplay. I've also done some of the usual format and
style improvement, mainly in IdToFileMap and ControlPanel.