Decouple RecentFileList from Preferences
RecentFileList required a Preferences object to provide a backing store,
but this was an unnecessarily heavy dependency since it only required a
getter for the constructor and a putter and remover for the save method.
It now takes functional interfaces as arguments instead of Preferences
objects, making unit testing and future reuse of the class easier.