Don't label points with invalid treatment levels
PuffinPlot can load files without treatment levels specified for the
steps; in this case it just displays them in order. However, if the
"label treatment steps" preference is active, this resulted in each
point on the Zplot and treatment step equal-area plot being labelled
with "NaN". PuffinPlot now omits the annotation if the formatted
treatment level is "NaN".