Warn user when jar cannot be saved to bundle, etc.
If "include jar file" is selected when creating a bundle, and if the
jar file cannot be copied (e.g. because PuffinPlot is not running from
a jar), PuffinPlot creates a jar-free bundle instead. Previously this
was done silently; now a warning is displayed to the user.

Additionally, modify the PuffinApp#errorDialog methods to wrap the
supplied text, thus avoiding ludicrously wide dialog boxes.