Core section end direction: allow missing values
Previously, calculation of a core section end direction required
directions to be defined for *all* samples near the end. This
requirement has now been relaxed: the end direction is calculated as the
mean of all the defined sample directions, missing values are ignored,
and an exception is only raised if there are *no* defined directions
near the end.