JR6 file loader: support default treatment type
PuffinPlot previously required a letter prefix for the treatment field
of an imported JR6 file in order to specify the treatment type. It turns
out that some JR6 files don't contain this prefix, and thus don't specify
the treatment type anywhere in the line.

The JR6 loader now accepts a default treatment type as an option, and
uses this for lines for which no treatment type is specified. PuffinApp
asks the user to specify this default via a dialog when importing a JR6
file.