Change Suite to use new FileLoader2 API
The Suite.readFiles methods now use the redesigned file loading API.
Lots of files touched in this commit in part because of a lot
of TwoGeeLoader.Protocol -> TwoGeeLoader2.Protocol replacements.
Also added unit tests for PmdLoader2, which I forgot to do when porting
PmdLoader.

Some more refactoring will be needed to clean up Suite.readFiles and
take full advantage of the new API, but it's fully functional in its
present state. I tested the desktop application by hand with all
supported filetypes; everything seems to be in working order.