Generate changelog in manual from CHANGES.md
The 1.03 changelog existed in duplicate: as the Markdown file CHANGES.md
in the repository root and as LaTeX source in the manual. I have now
improved the manual build script to automatically generate LaTeX from
CHANGES.md using Pandoc, avoiding the need to manually synchronize the two
changelogs.

With this commit, the 1.4 changelog also becomes a part of the manual,
since it is already present in CHANGES.md.