Sensor length warning: only for continuous data
The 2G file loader should produce a warning message if continuous
Cartesian data is loaded with unset sensor lengths. When this check was
moved from Suite to TwoGeeLoader in commit 1d3ddc07, the check for a
continuous suite was omitted, so the warning was being given even for
discrete data. Bug now fixed (and unit test added), and warning is issued
only when appropriate.