Warn user if 2G file has few or no valid headers
PuffinPlot makes a best effort with partial data. In the case of the
extremely variable 2G format, this means that it will proceed with file
loading even if not a single column header is recognized. However, this
circumstance probably indicates that something is wrong with the file,
so PuffinPlot now issues a warning if three or fewer column headers
are recognized in a 2G file.