Improve axis labels on demagnetization plot
When plotting remanence and susceptibility on the same demagnetization
plot, the tick and label were omitted at 0 to make space for a legend
point indicating which point style was associated with which data set.
However, this was also done unnecessarily when the second (susceptibility)
y axis was absent; additionally, the tick and label at 0 were omitted
on the x axis.

Improvements: tick and label at 0 now always present on the x axis. On the
y axes, the ticks and labels at 0 are only replaced with legend points
when both axes are present. On a remanence-only plot, the tick and label
at 0 are shown instead.