First steps in converting manual to Asciidoc
Added a very rough first Asciidoc version of the manual (converted
automatically from LaTeX using Pandoc). Also made a minor syntax change to
the LaTeX manual to make it comprehensible to Pandoc. The Asciidoc manual
is not yet used at all in the build process, and will need a lot of work
before it can serve as a replacement for the LaTeX version.