Change quotation marks in asciidoc manual source
The asciidoc manual previously encoded curved quotation marks and
apostrophes by using the appropriate UTF-8 characters ‘ and ’ in the
source. While this produces the desired output, it is not the standard
method in asciidoc. I have now changed to the usual asciidoc encodings of
'` and `' for quotation marks and plain ' for apostrophes, which are
replaced as appropriate with ‘ and ’ when the source is processed.