Use asciidoctor-bibtex in Asciidoc manual
Use the asciidoctor-bibtex ruby gem for automatic reference handling in
the Asciidoc manual. This involves using the gem-maven-plugin to fetch
asciidoctor-bibtex and its dependencies from the torquebox.org
gem-proxying Maven repository. The manual source itself has not yet been
updated to use asciidoctor-bibtex, with the exception of one citation and
a corresponding one-item bibliography as a check that it's working
correctly.