Switch from git to jgit for getting commit info
Previously, the build process executed git to obtain some information on
the current commit for later use at runtime. This introduced a dependency
on the git executable, compromising the portability of the build. The same
information is now fetched using jgit and a jruby script, removing what I
believe to be the last non-Maven-fetchable dependency in the build
process.