Make build possible in absence of git repository
The jruby/jgit scripts used in the build assumed the presence of a git
repository, and broke the build if none was present. They now check that a
repository is present before trying to read information from it. If no
repository is present, PuffinPlot and the manual use fallback values for
version and date info.